requirements.txt
opencv-python
scipy
numpy
pillow
pyyaml
pandas
natsort
plyfile
